1.2.0 Add non-static inTransaction to Persistence and PersistenceFactory
The inTransaction on Persistence will use the instance in the transaction.
The inTransaction on PersistenceFactory will use the factory to create a persistence, us that for the transaction, then close that persistence.
Also corrected error handling in the event that the rollback fails so that the suppressed error will not be lost. (Ran into difficulty with the h2 driver not supporting transactions in the way I expected it to. I may have to use another driver for my examples)